pyproject.toml
[tool.poetry]
name = "Discord-bot"
version = "0.1.0"
description = ""
authors = ["FrancoGiachetta27 <fgiachetta@frba.utn.edu.ar>"]

[tool.poetry.dependencies]
python = "^3.10"
discord = "^1.0.1"
youtube-dl = "^2021.12.17"
async-timeout = "^3.0.1"
PyNaCl = "^1.4.0"
python-dotenv = "^0.20.0"
spotipy = "^2.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
